--- src/app/_chat_mvp_2/layout.tsx
@@ -1,9 +1,0 @@
-import AuthLayout from "@/components/AuthLayout"
-
-export default function ChatLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return <AuthLayout>{children}</AuthLayout>
-}
