+++ src/app/auth/signin/page.tsx
@@ -1,189 +1,20 @@
-'use client'
-
-import { signIn } from "next-auth/react"
-import { useState } from "react"
-import { useRouter, useSearchParams } from "next/navigation"
-import Logo from "@/components/Logo"
+import SignInClient from "@/components/SignInClient"
 import { getSiteConfig } from "@/lib/site-config"
 import { getSafeCallbackUrl } from "@/lib/callback-utils"
 
-export default function SignIn() {
-  const [isLoading, setIsLoading] = useState(false)
-  const [username, setUsername] = useState("")
-  const [password, setPassword] = useState("")
-  const [error, setError] = useState("")
-  const router = useRouter()
-  const searchParams = useSearchParams()
-  const next = getSafeCallbackUrl(searchParams.get('next'))
+interface SignInPageProps {
+  searchParams: Promise<{ [key: string]: string | string[] | undefined }>
+}
+
+export default async function SignIn({ searchParams }: SignInPageProps) {
+  const params = await searchParams
+  const next = getSafeCallbackUrl(params.next as string) || "/chat"
   const siteConfig = getSiteConfig()
 
-  const handleCredentialsSignIn = async (e: React.FormEvent) => {
-    e.preventDefault()
-    setIsLoading(true)
-    setError("")
-
-    try {
-      const result = await signIn('credentials', {
-        username,
-        password,
-        redirect: false,
-      })
-
-      if (result?.error) {
-        setError("Invalid username or password")
-      } else {
-        router.push(next)
-      }
-    } catch (err) {
-      console.error('Sign in error:', err)
-      setError("An error occurred. Please try again.")
-    } finally {
-      setIsLoading(false)
-    }
-  }
-
-  const handleGoogleSignIn = async () => {
-    setIsLoading(true)
-    await signIn('google', { callbackUrl: next })
-    setIsLoading(false)
-  }
-
-  const handleMockSignIn = () => {
-    // For development, just redirect to the callback URL (mock auth is handled server-side)
-    router.push(next)
-  }
-
   return (
-    <div className="min-h-screen flex items-center justify-center py-12 px-4 sm:px-6 lg:px-8" style={{ backgroundColor: 'var(--background)' }}>
-      <div className="max-w-md w-full space-y-8">
-        <div>
-          <div className="mx-auto">
-            <Logo className="justify-center" />
-          </div>
-          <h2 className="mt-6 text-center text-3xl font-extrabold" style={{ color: 'var(--primary)' }}>
-            Welcome to {siteConfig.title}
-          </h2>
-          <p className="mt-2 text-center text-sm text-gray-600">
-            Sign in to continue to your conversations
-            <br />
-            <span className="text-xs text-gray-500 mt-1 block">
-              Demo: username &quot;user&quot;, password &quot;pass&quot;
-            </span>
-          </p>
-        </div>
-
-        <form className="mt-8 space-y-6" onSubmit={handleCredentialsSignIn}>
-          <div>
-            <label htmlFor="username" className="sr-only">
-              Username
-            </label>
-            <input
-              id="username"
-              name="username"
-              type="text"
-              autoComplete="username"
-              required
-              value={username}
-              onChange={(e) => setUsername(e.target.value)}
-              className="relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 rounded-t-md focus:outline-none focus:ring-2 focus:border-transparent focus:z-10 sm:text-sm"
-              style={{ 
-                backgroundColor: 'var(--background)',
-                color: 'var(--foreground)',
-                borderColor: 'var(--primary)'
-              }}
-              placeholder="Username"
-            />
-          </div>
-          <div>
-            <label htmlFor="password" className="sr-only">
-              Password
-            </label>
-            <input
-              id="password"
-              name="password"
-              type="password"
-              autoComplete="current-password"
-              required
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-              className="relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 rounded-b-md focus:outline-none focus:ring-2 focus:border-transparent focus:z-10 sm:text-sm"
-              style={{ 
-                backgroundColor: 'var(--background)',
-                color: 'var(--foreground)',
-                borderColor: 'var(--primary)'
-              }}
-              placeholder="Password"
-            />
-          </div>
-
-          {error && (
-            <div className="text-red-600 text-sm text-center">
-              {error}
-            </div>
-          )}
-
-          <div>
-            <button
-              type="submit"
-              disabled={isLoading}
-              className="group relative w-full flex justify-center py-2 px-4 border border-transparent text-sm font-medium rounded-md text-white hover:opacity-80 focus:outline-none focus:ring-2 focus:ring-offset-2 disabled:opacity-50"
-              style={{ backgroundColor: 'var(--primary)' }}
-            >
-              {isLoading ? (
-                <svg className="animate-spin -ml-1 mr-3 h-5 w-5 text-white" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
-                  <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
-                  <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
-                </svg>
-              ) : null}
-              {isLoading ? 'Signing in...' : 'Sign in'}
-            </button>
-          </div>
-        </form>
-
-        <div className="mt-6">
-          <div className="relative">
-            <div className="absolute inset-0 flex items-center">
-              <div className="w-full border-t border-gray-300" />
-            </div>
-            <div className="relative flex justify-center text-sm">
-              <span className="px-2 text-gray-500" style={{ backgroundColor: 'var(--background)' }}>Or continue with</span>
-            </div>
-          </div>
-
-          <div className="mt-6 space-y-3">
-            <button
-              onClick={handleMockSignIn}
-              className="group relative w-full flex justify-center py-2 px-4 border border-gray-300 text-sm font-medium rounded-md hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2"
-              style={{ 
-                backgroundColor: 'var(--background)',
-                color: 'var(--foreground)'
-              }}
-            >
-              <svg className="w-5 h-5 mr-2" fill="currentColor" viewBox="0 0 24 24">
-                <path d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm-2 15l-5-5 1.41-1.41L10 14.17l7.59-7.59L19 8l-9 9z"/>
-              </svg>
-              Continue with Mock User (Development)
-            </button>
-            <button
-              onClick={handleGoogleSignIn}
-              disabled={isLoading}
-              className="group relative w-full flex justify-center py-2 px-4 border border-gray-300 text-sm font-medium rounded-md hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 disabled:opacity-50"
-              style={{ 
-                backgroundColor: 'var(--background)',
-                color: 'var(--foreground)'
-              }}
-            >
-              <svg className="w-5 h-5 mr-2" viewBox="0 0 24 24">
-                <path fill="currentColor" d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"/>
-                <path fill="currentColor" d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"/>
-                <path fill="currentColor" d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"/>
-                <path fill="currentColor" d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"/>
-              </svg>
-              Sign in with Google
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
+    <SignInClient 
+      callbackUrl={next}
+      siteConfig={siteConfig}
+    />
   )
 }
